nt_copper_prospectivity_repo/nt-copper-prospectivity-kobold-style/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nt-cu-prospectivity"
version = "0.1.0"
description = "style PU ensemble prospectivity mapping for Northern Territory (Cu)."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Tran Phong"}]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "pyarrow>=16",
  "geopandas>=1.0",
  "shapely>=2.0",
  "pyproj>=3.6",
  "rasterio>=1.3",
  "rioxarray>=0.15",
  "xarray>=2024.3",
  "scikit-learn>=1.5",
  "joblib>=1.3",
  "tqdm>=4.66",
  "matplotlib>=3.8",
  "typer>=0.12",
  "rich>=13.7",
  "pydantic>=2.7",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.2","ruff>=0.5"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E","F","I","UP","B","SIM","RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
